+++ app/helpers/application_helper.rb
@@ -1,19 +1,4 @@
 module ApplicationHelper
-  OVERRIDE_BROWSE_PAGES = {
-    '/browse/driving/driving-licences' => [
-      {
-        title: 'Learn to drive a car: step by step',
-        base_path: '/services/learn-to-drive-a-car'
-      }
-    ],
-    '/browse/driving/learning-to-drive' => [
-      {
-        title: 'Learn to drive a car: step by step',
-        base_path: '/services/learn-to-drive-a-car'
-      }
-    ]
-  }
-
   def set_tasklist(tasklist)
     session[:tasklist] = tasklist
   end
@@ -45,24 +30,6 @@ module ApplicationHelper
       links = task["task_items"].map { |task| task["base_path"] }
       links.include? base_path
     end
-  end
-
-  def browsing_in_top_level_page?(top_level_page)
-    request.path.starts_with?(top_level_page.base_path)
-  end
-
-  def browsing_in_second_level_page?(section)
-    request.path.starts_with?(section.base_path)
-  end
-
-  def override_browse_page?(params)
-    path = "/browse/driving/#{params[:second_level_slug]}"
-    OVERRIDE_BROWSE_PAGES.keys.include?(path)
-  end
-
-  def override_browse_page_with(params)
-    path = "/browse/driving/#{params[:second_level_slug]}"
-    OVERRIDE_BROWSE_PAGES[path]
   end
 
   def render_popular_list?
