+++ app/views/second_level_browse_page/_links.html.erb
@@ -10,7 +10,7 @@
 
     <ul>
       <% if override_browse_page?(params) && index.zero? %>
-        <% override_browse_page_with(params).each do |link|%>
+        <% override_browse_page_with(params).each do |link| %>
         <li>
           <a href="<%= link[:base_path] %>"><%= link[:title]%></a>
         </li>
@@ -23,7 +23,17 @@
         <% end %>
       <% else %>
         <% list.contents.each_with_index do |list_item, list_index| %>
-          <li><%= link_to list_item.title, list_item.base_path %></li>
+          <% if list_index.zero? && override_divorce_browse_page?(list.title) %>
+            <% divorce_service_browse_links.each do |link| %>
+            <li>
+              <%= link_to link[:title], link[:base_path] %>
+            </li>
+            <% end %>
+          <% end %>
+
+          <% unless disable_browse_link?(list_item.base_path)%>
+            <li><%= link_to list_item.title, list_item.base_path %></li>
+          <% end%>
         <% end %>
       <% end %>
     </ul>
